Refactor Serializer logic to use organisers
