+++ kit/libs/kit-json-api/app/services/kit/json_api/services/serializer/query.rb
@@ -1,0 +1,33 @@
+module Kit::JsonApi::Services::Serializer::Query
+  include Kit::Contract
+  Ct = Kit::JsonApi::Contracts
+
+  before Ct::Hash[query_node: Ct::QueryNode]
+  after  Ct::Result[document: Ct::Document]
+  # Entry point to serialize a query AST
+  def self.serialize_query(query_node:)
+    Kit::Organizer.call({
+      list: [
+        self.method(:create_document),
+        Kit::JsonApi::Services::Serializer::QueryNode.method(:serialize_query_node),
+      ],
+      ctx: { query_node: query_node, },
+    })
+  end
+
+  after  Ct::Result[document: Ct::Document]
+  # Create a Document object that contains the json response and various caches
+  def self.create_document
+    document = Kit::JsonApi::Types::Document[{
+      cache:    {},
+      included: {},
+      response: {
+        data:     [],
+        included: [],
+      },
+    }]
+
+    [:ok, document: document]
+  end
+
+end
